[BE] feat: 커스텀 키보드 생성, 상세조회 api 추가

See merge request s12-final/S12P31A703!73

+++ server/spring/src/main/java/com/ssafy/tadak/spring/common/resolver/UserInfoArgumentResolver.java
@@ -34,6 +34,6 @@ public class UserInfoArgumentResolver implements HandlerMethodArgumentResolver {
             throw new UnauthorizedException(new ErrorCode("B4015", "인증되지 않은 요청입니다."));
         }
 
-        return new UserInfo(Integer.parseInt(userId), nickname, role);
+        return new UserInfo(Long.parseLong(userId), nickname, role);
     }
 }
